refactor: storeInfoAtom /shared로 이동
