--- src/features/upload-image/atoms/uploadImageAtoms.ts
@@ -1,4 +1,0 @@
-import { StoreInfoT } from '@/entities'
-import { atom } from 'jotai'
-
-export const storeInfoAtom = atom<StoreInfoT>()
