+++ src/features/upload-image/ui/UploadImageScreen.tsx
@@ -6,15 +6,13 @@ import { useRouter } from 'next/navigation'
 import { ButtonMobile, Icon, TopBar } from '@eolluga/eolluga-ui'
 import Link from 'next/link'
 import Image from 'next/image'
-import { ToastMessage, sendRNFunction } from '@/shared'
+import { ToastMessage, sendRNFunction, storeInfoAtom } from '@/shared'
 import { OrbitProgress } from 'react-loading-indicators'
-import { useGetStoreInfo } from '@/entities'
 import storeDefaultImage from '@public/image/store_default_image.png'
+import { useAtom } from 'jotai'
 import { usePutStoreImage } from '../model/usePutStoreImage'
 import { ImageUploadResultT } from '../types/imageUploadType'
 import { useGetStoreImage } from '../model/useGetStoreImage'
-import { useAtom } from 'jotai'
-import { storeInfoAtom } from '../atoms/uploadImageAtoms'
 
 interface ImageUploadScreenProps {
   page: 'home' | 'join'
